fix: rename to lib/validators and add lint guard
