+++ app/(auth)/register/components/RegisterForm.test.tsx
@@ -4,7 +4,7 @@ import userEvent from "@testing-library/user-event";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
 import { RegisterForm } from "@root/app/(auth)/register/components/RegisterForm";
-import { validateAccount } from "@lib/client/validationSchemas";
+import { validateAccount } from "@lib/validation/validationSchemas";
 import { useTranslation } from "@i18n";
 
 import { createRouterStub, createTranslationStub } from "../../../../test/helpers/client";
@@ -30,7 +30,7 @@ vi.mock("@i18n/client", () => ({
   LANGUAGE_COOKIE_NAME: "i18next",
 }));
 
-vi.mock("@lib/client/validationSchemas", () => ({
+vi.mock("@lib/validation/validationSchemas", () => ({
   validateAccount: vi.fn(),
 }));
