+++ app/(auth)/register/password/components/SetRegisterPasswordForm.tsx
@@ -7,7 +7,7 @@ import { PasswordComplexitySettings } from "@zitadel/proto/zitadel/settings/v2/p
 /*--------------------------------------------*
  * Internal Aliases
  *--------------------------------------------*/
-import { validateAccount } from "@lib/client/validationSchemas";
+import { validateAccount } from "@lib/validation/validationSchemas";
 import { useTranslation } from "@i18n/client";
 import { PasswordValidationForm } from "@components/auth/password-validation/PasswordValidationForm";
 import { Alert, ErrorStatus } from "@components/ui/form";
