+++ app/account/components/PersonalDetails.tsx
@@ -7,12 +7,12 @@ import { useActionState, useEffect, useRef, useState } from "react";
 import { flushSync } from "react-dom";
 import { useTranslation } from "react-i18next";
 
+import { cn } from "@lib/utils";
 /*--------------------------------------------*
  * Internal Aliases
  *--------------------------------------------*/
-import { validatePersonalDetails } from "@lib/client/validationSchemas";
-import { getError, hasError } from "@lib/client/validators";
-import { cn } from "@lib/utils";
+import { validatePersonalDetails } from "@lib/validation/validationSchemas";
+import { getError, hasError } from "@lib/validation/validators";
 import { Button } from "@components/ui/button/Button";
 import { SubmitButtonAction } from "@components/ui/button/SubmitButton";
 import { Label, TextInput } from "@components/ui/form";
